player_data_service/pyproject.toml
[project]
name = "player_data_service"
version = "0.1.0"
requires-python = ">= 3.11"
readme = "README.md"
authors = [{ name = "Zachary Foteff", email = "zfoteff@protonmail.com" }]
maintainers = [{ name = "Zachary Foteff", email = "zfoteff@protonmail.com" }]
dependencies = [
    "annotated-types==0.6.0",
    "anyio==4.3.0",
    "black==22.3.0",
    "certifi==2024.2.2",
    "click==8.1.7",
    "fastapi==0.110.0",
    "flake8==4.0.1",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httptools==0.6.1",
    "httpx==0.27.0",
    "idna==3.6",
    "iniconfig==2.0.0",
    "isort==5.10.1",
    "mccabe==0.6.1",
    "mypy-extensions==1.0.0",
    "mysql-connector-python==8.4.0",
    "packaging==24.0",
    "pathspec==0.12.1",
    "platformdirs==4.2.1",
    "pluggy==1.5.0",
    "pycodestyle==2.8.0",
    "pydantic==2.7.1",
    "pydantic_core==2.18.2",
    "pyflakes==2.4.0",
    "pytest==8.1.1",
    "python-dotenv==1.0.1",
    "sniffio==1.3.1",
    "starlette==0.36.3",
    "typing_extensions==4.10.0",
    "uvicorn==0.29.0",
    "uvloop==0.19.0",
    "watchfiles==0.21.0",
    "websockets==12.0",
]

[project.urls]
Repository = "https://github.com/zfoteff/clackamas-lacrosse-application/tree/develop/src/player_data_service"
Issues = "https://github.com/zfoteff/clackamas-lacrosse-application/issues"
